backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
redis
python-dotenv